win32 sigterm impl

+++ src/CompilerProgramOptions.cpp
@@ -25,7 +25,7 @@ Validate mode checks for errors and returns the validation result as json object
         ;
     po::positional_options_description p;
     p.add(ARG_INPUT, -1);
-    po::store(po::command_line_parser(argc, argv).
+    po::store(po::command_line_parser((int)argc, argv).
         options(optionsDescription).positional(p).run(), variables);
     po::notify(variables);
 }
